src/deployml/terraform/modules/teardown/cloud/gcp/cloud_function/requirements.txt
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0

